fetch currentuser data from api using axios
