--- src/redux/house/userHouseSlice.js
@@ -1,32 +1,0 @@
-/* eslint-disable no-param-reassign */
-import { createSlice } from '@reduxjs/toolkit';
-import axios from 'axios';
-
-const initialState = {
-    houses: [],   // empty array to initalize house
-    isLoading: false,
-    error: null
-};
-
-const userHouseSlice = createSlice({
-    name: 'userHouse',
-    initialState,
-    reducers: {
-        fetchUserHouseStart: (state) => {
-            state.isLoading = true;
-            state.error = null;
-        },
-        fetchUserHouseSuccess: (state, action) => {
-            state.houses = action.payload;
-            state.isLoading = false;
-            state.error = null;
-        },
-        fetchUserHouseFail: (state, house) => {
-            state.isLoading = false;
-            state.error = action.payload;
-
-        },
-    },
-});
-
-export const { fetchUserHouseStart, fetchUserHouseSuccess, fetchUserHouseFail } = userHouseSlice.actions;
